+++ TaskTracker.API/TaskTracker.API.csproj
@@ -17,6 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\TaskTracker.Application\TaskTracker.Application.csproj" />
+    <ProjectReference Include="..\TaskTracker.Database\TaskTracker.Database.csproj" />
     <ProjectReference Include="..\TaskTracker.Infrastructure\TaskTracker.Infrastructure.csproj" />
     <ProjectReference Include="..\TaskTracker.Persistence\TaskTracker.Persistence.csproj" />
   </ItemGroup>
